Changed to show meals for logged in store | Admin Store view with Store drop down to add new | few other small fixes

+++ database/migrations/2018_10_13_031915_create_meals_table.php
@@ -20,7 +20,7 @@ class CreateMealsTable extends Migration
             $table->string('title');
             $table->string('category');
             $table->text('description');
-            $table->integer('price');
+            $table->decimal('price', 4, 2);
             $table->timestamps();
         });
     }
